requirements.txt
torch==2.0.0
torchvision==0.15.0
tqdm==4.64.1
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.0
matplotlib==3.7.1
seaborn==0.12.2
Pillow==9.4.0
opencv-python==4.7.0.72
albumentations==1.3.0
termcolor==2.2.0
timm==0.9.2
